+++ client/src/components/Form/log_in.jsx
@@ -1,7 +1,6 @@
 import React, {useState} from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import axios from "axios";
-import LogOut from '../LogOut/LogOut';
 
 
 export default function LogIn() {
@@ -109,7 +108,6 @@ export default function LogIn() {
           <Link to='/home'>
            O entra como invitado
           </Link>
-          <LogOut />
         </div>
   )
 }
