--- client/src/components/LogOut/LogOut.css
@@ -1,24 +1,0 @@
-.btn_log_out{
-    background-color: rgb(207, 42, 17);
-    margin: 10px;    
-    font-size: 20px;
-    padding: 5px;
-    text-align: center;
-    text-transform: uppercase;
-    transition: 0.5s;
-    background-size: 200% auto;
-    color: #FFF;
-    box-shadow: 0 0 20px #eee;
-    border-radius: 10px;    
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.807), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(.25,.8,.25,1);
-    cursor: pointer;
-     border-radius: 5px;
- }
-
- .btn_log_out:hover{
- box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
-    margin: 8px 10px 12px;
-    font-size: 25px;
-    background: -webkit-linear-gradient(rgb(255, 0, 0), rgb(249, 249, 249));
- }
